requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Django==4.2.16
Flask==3.0.3
fonttools==4.53.1
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mysql-connector-python==9.0.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
psycopg2-binary==2.9.10
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.0.4
XlsxWriter==3.2.0
zipp==3.20.1
